engine/requirements.txt
argcomplete==1.8.2
beautifulsoup4==4.5.3
chardet==2.3.0
lxml==4.2.4
mongoengine==0.15.3
numpy==1.15.1
Pillow==5.2.0
pymongo==3.7.1
pytz==2018.5
scikit-learn==0.19.2
scipy==1.1.0
XlsxWriter==1.0.9
amqp==2.4.2
autopep8==1.3.5
billiard==3.5.0.5
celery==4.2.1
certifi==2019.3.9
cymem==1.31.2
cytoolz==0.8.2
dill==0.2.7.1
django==2.2
django-cors-headers==2.5.2
django-filter==1.1.0
django-mongodb-engine==0.6.0
django-rest-framework-mongoengine==3.3.1
djangorestframework
djangotoolbox==1.8.0
docx2txt==0.6
EbookLib==0.15
future==0.16.0
gevent==1.3b2
greenlet==0.4.13
gunicorn==19.6.0
idna==2.8
kombu==4.5.0
Markdown==2.6.11
msgpack-numpy==0.4.1
msgpack-python==0.5.6
murmurhash==0.28.0
mysql-connector==2.1.7
nltk==3.3
pathlib==1.0.1
pdfminer.six==20170720
plac==0.9.6
pocketsphinx==0.1.3
preshed==1.0.0
pycodestyle==2.4.0
pycryptodome==3.6.1
pyocclient==0.4
PyPDF2==1.26.0
python-dateutil==2.7.5
python-pptx==0.6.5
pyzmq==17.0.0
redis==3.2.0
regex==2017.4.5
requests==2.21.0
six==1.10.0
spacy==2.0.11
SpeechRecognition==3.6.3
termcolor==1.1.0
textblob==0.15.2
textract==1.6.1
thinc==6.10.2
toolz==0.9.0
tqdm==4.23.2
ujson==1.35
urllib3==1.22
vine==1.3.0
wrapt==1.10.11
xlrd==1.0.0
#xx-ent-wiki-sm==2.0.0
